requirements.txt
django==4.1.4
djangorestframework==3.14.0
psycopg2==2.9.5
redis==4.4.0
django-celery-beat==2.4.0
django-celery-results==2.4.0
python-decouple==3.6
requests==2.28.1
